Ajout de la fonctionnalité d'authentification

+++ admin/delete-apprenant.php
@@ -1,4 +1,11 @@
 <?php
+    require "../function.php";
+    if (!isConnected()){
+        header("location:login.php");
+        die;
+    
+    }
+    
     $appr_id=$_GET['id'];
     include_once "../db-connect.php";
     $sql= "DELETE FROM apprenants where id=$appr_id";
